requirements.txt
Flask==3.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
cloudinary==1.38.0
python-dotenv==1.0.0